requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
opencv-python>=4.5.0
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.6.0
scikit-learn>=0.24.2

# Computer Vision
ultralytics>=8.0.0  # YOLOv8
albumentations>=1.0.0

# Robotics Control
pyserial>=3.5
RPi.GPIO>=0.7.0  # For Raspberry Pi GPIO
gpiozero>=1.6.2

# Web Dashboard
streamlit>=1.0.0
plotly>=5.3.0
dash>=2.0.0

# API
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.7.0

# Blockchain
web3>=5.24.0
eth-account>=0.5.6

# Monitoring & Logging
prometheus-client>=0.11.0
python-dotenv>=0.19.0
loguru>=0.5.3

# Testing
pytest>=6.2.5
pytest-cov>=2.12.0
pytest-asyncio>=0.15.0

# Development
black>=21.7b0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0

# Documentation
sphinx>=4.1.0
sphinx-rtd-theme>=0.5.2
nbsphinx>=0.8.7

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Fairness and bias detection
aif360>=0.4.0
fairlearn>=0.7.0

# Privacy
diffprivlib>=0.5.0
opacus>=1.0.0

# Explainability
shap>=0.40.0
lime>=0.2.0
interpret>=0.2.7

# Storage and serialization
pyyaml>=6.0.0
joblib>=1.1.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Monitoring and logging
python-json-logger>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Type checking
types-PyYAML>=6.0.0
types-python-dateutil>=2.8.0

# Development tools
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0

# Additional dependencies
tensorflow-privacy>=0.8.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
dash>=2.0.0
streamlit>=1.8.0
mlflow>=2.0.0
wandb>=0.12.0
sphinx>=4.5.0
mkdocs>=1.4.0
mkdocs-material>=8.0.0
pytest>=7.0.0
pytest-cov>=3.0.0
click>=8.0.0
typer>=0.4.0
tqdm>=4.62.0
python-dotenv>=0.19.0
pyyaml>=6.0.0
onnx>=1.14.0
onnxruntime-gpu>=1.16.0
tensorflow-hub>=0.14.0
albumentations>=1.3.1
ros-humble-ros-base>=0.4.0
pymodbus>=3.5.0
python-etcd>=0.4.5
matplotlib>=3.8.0
seaborn>=0.13.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
python-jose>=3.3.0
passlib>=1.7.4
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
redis>=5.0.0
influxdb-client>=1.36.0
web3>=6.11.0
eth-account>=0.9.0
eth-typing>=3.5.0
python-logging-loki>=0.3.1
grafana-api>=1.0.3
pre-commit>=2.17.0
safety-gym>=0.2.0
scipy>=1.7.0 